requirements.txt
azure-cognitiveservices-speech=1.21.0
